fghgfhgfhfg

+++ .ht/API/deletePic.php
@@ -22,14 +22,26 @@ class deletePic{
             showMessage(false, 'Error 1');
         }elseif (!isset($data['whois']) || empty($data['whois'])) {
             showMessage(false, 'Error 2');
-        }elseif (!isset($data['extension']) || empty($data['extension'])) {
-            showMessage(false, 'Error 3');
         }elseif (!isset($data['personID']) || empty($data['personID'])) {
             showMessage(false, 'Error 4');
-        }elseif (!$this->checkUserE()) {
+        }elseif (!isset($data['purpose']) || empty($data['purpose'])) {
             showMessage(false, 'Error 5');
-        }else{
-            if ($data['whois'] == 'user') {
+        }elseif (!$this->checkUserE()) {
+            showMessage(false, 'Error 6');
+        }elseif ($data['purpose'] == 'visibility') {
+            if (!isset($data['value']) || empty($data['value'])) {
+                showMessage(false, 'Error 6');
+            }elseif ($data['whois'] == 'user') {
+                $this->changeByUser();
+            }elseif ($data['whois'] == 'admin') {
+                $this->changeByAdmin();
+            }else{
+                showMessage(false, 'Error 5');
+            }
+        }elseif ($data['purpose'] == 'delete') {
+            if (!isset($data['extension']) || empty($data['extension'])) {
+                showMessage(false, 'Error 3');
+            }elseif ($data['whois'] == 'user') {
                 $this->deleteByUser();
             }elseif ($data['whois'] == 'admin') {
                 $this->deleteByAdmin();
@@ -39,8 +51,50 @@ class deletePic{
         }
     }
 
+
+    private function changeByUser(){
+        $data = json_decode(file_get_contents('php://input'), true);
+        $eid = $data['personID'];
+        $personID = $this->AUTH->decrypt($eid); 
+        $username = $this->userData->getOtherData('personID', $personID)['username'];
+        $imgID = $data['imgID'];
+        $value = $data['value'];
+        if ($personID != $_SESSION['LOGGED_USER']) {
+            showMessage(false, 'Error 7');
+        }else{
+            $sql = "UPDATE uploads set access = '$value' WHERE uploadID = '$imgID' and username = '$username'"; 
+            $result = mysqli_query($this->DB, $sql);
+            if($result){
+                showMessage(true, 'changed');
+            }else{
+                showMessage(false, "Can't changed");
+            }
+        }
+    }
+
+    private function changeByAdmin(){
+        $data = json_decode(file_get_contents('php://input'), true);
+        $eid = $data['personID'];
+        $personID = $this->AUTH->decrypt($eid); 
+        $username = $this->userData->getOtherData('personID', $personID)['username'];
+        $imgID = $data['imgID'];
+        $value = $data['value'];
+        if ($this->userData->getSelfDetails()['userType'] != 'Admin') {
+            showMessage(false, 'Error 7');
+        }elseif($value != 'everyone' || $value != 'self' || $value != 'followers'){
+            $value = 'none';
+        }else{
+            $sql = "UPDATE uploads set access = '$value' WHERE uploadID = '$imgID' and username = '$username'"; 
+            $result = mysqli_query($this->DB, $sql);
+            if($result){
+                showMessage(true, 'changed');
+            }else{
+                showMessage(false, "Can't changed");
+            }
+        }
+    }
+
     private function deleteByUser(){
-        
         $data = json_decode(file_get_contents('php://input'), true);
         $eid = $data['personID'];
         $personID = $this->AUTH->decrypt($eid); 
@@ -49,7 +103,7 @@ class deletePic{
         $ext = $data['extension'];
         $path = $this->_DOCROOT.'/fastreedusercontent/photos/'.$username.'/'.$imgID.$ext;
         if ($personID != $_SESSION['LOGGED_USER']) {
-            showMessage(false, 'Error 6');
+            showMessage(false, 'Error 7');
         }elseif (file_exists($path)) {
             if (unlink($path)) {
                 $sql = "DELETE FROM uploads WHERE personID = '$personID' and uploadID = '$imgID'";
